directory url pages moved integration in directory moved assets
