+++ astro.config.mjs
@@ -9,7 +9,7 @@ import {remarkPanzoom} from './src/libs/remark-panzoom'
 import {remarkGallery} from './src/libs/remark-gallery'
 import {config} from './config'
 import {replaceFiledir} from './src/libs/vite-plugin-filedir'
-import {generate_menu} from './src/libs/integration-generate-menu'
+import {generate_menu} from './integrations/nav_menu/integration-generate-menu'
 
 const default_options = {
   markdown:{
