+++ data/blog/panzoom/readme.mdx
@@ -35,7 +35,3 @@ SVG Files can be used in markdown too, they'll be passed through as is.
 <svg viewBox="0 0 100 100" width="60" height="60" fill="#00000000" xmlns="http://www.w3.org/2000/svg">
     <path d="M 20,10 L 70,50 L 20,90" stroke-width="20px" stroke="#d0d0d0" stroke-linecap="round" stroke-linejoin="round"></path>
 </svg>
-
-test invisible diagram
-
-![gal](./gallery-long_diag.svg)
