+++ src/components/icon.astro
@@ -1,6 +1,6 @@
 ---
 
-import githubImg from '@/img/github-dark.png'
+import githubImg from '@/assets/github-dark.png'
 export interface Props {
   filename: string;
 }
