+++ src/components/svgicons.astro
@@ -1,12 +1,14 @@
 ---
 //https://ellodave.dev/blog/article/using-svgs-as-astro-components-and-inline-css/
 import {existsSync,readFileSync} from 'fs'
+import { config } from '@/config';
+import {join} from 'node:path'
 export interface Props {
   filename: string;
 }
 
 const { filename } = Astro.props as Props;
-const filepath = `./src/svg/${filename}.svg`
+const filepath = join(config.rootdir,`src/assets/${filename}.svg`)
 const innerHTML = existsSync(filepath)?readFileSync(filepath).toString():"<svg/>";
 ---
