+++ src/components/svgurl.astro
@@ -2,7 +2,7 @@
 //https://ellodave.dev/blog/article/using-svgs-as-astro-components-and-inline-css/
 import {join} from 'node:path'
 import {config} from '../../config'
-import {remove_base} from '../layout/menu_utils'
+import {remove_base} fro../../integrations/nav_menu/menu_utilss/nav_menu/menu_utils'
 import {promises as fs} from 'node:fs';
 export interface Props {
   url: string;
